requirements.txt
eventlet==0.25.1
psycopg2==2.8.4
python-socketio==4.4.0
